[56] Fix Edge semantic DnD with Metaclass target
In case of Metaclass target of edge, its representation is not the
representation of its container which is the UML metamodel. The
representation is the representation of the element import which refer
to this metaclass. So DnD of edge with Metaclass as target create now
the representation of Metaclass in the element import graphical
container.

Bug: https://github.com/PapyrusSirius/papyrus-web/issues/346
Bug: https://gitlab.eclipse.org/eclipse/papyrus/org.eclipse.papyrus-web/-/issues/56
Signed-off-by: Jessy Mallet <jessy.mallet@obeo.fr>