[215] Implement of a new default explorer tree description for UML
This first implementation make all the plumbing to provide a Tree
Description that can be used in the Explorer view.

The default implementation basically use the same structure as the
preview tree with the following exceptions:

* The EAnnoation are not longer displayed
* The Steretoype applications are not longer displayed
* The read only resources are no longer available if the filter "Hide
read only resources" is activated. In the previous implementation they
were hidden but still computed.
* Abstract elements are displayed in italic
* Static element are underlined
* ImportedElement and ImportedPackage should the imported element as a
child
* The label of Element with stereotypes are prefixed with the name of
the applied stereotypes

This commit also add the concept of read only elements.

Bug: https://gitlab.eclipse.org/eclipse/papyrus/org.eclipse.papyrus-web/-/issues/215
Signed-off-by: Arthur Daussy <arthur.daussy@obeo.fr>